support touch devices (show tooltip on focus):
Touch device support was almost there, as tooltip was already
shown on "focusin" events. The missing part was that nothing was
done on "focusout". So in the case of touch-only devices,
without mouse* events, but with focus* events, you could end up
having a tooltip that stayed visible, until something else was focused.

Now we make sure to hide the tooltip on focus out.